Integrate login flow
