+++ src/pages/Profile/Expert/Expert.css
@@ -74,9 +74,10 @@ button[aria-disabled='true'] {
   transform: scale(0.96);
 }
 
-.scrollable {
-  overflow-y: auto;
-  overflow-x: hidden;
+.ornament {
+  position: sticky;
+  bottom: 0;
+  z-index: 2;
 }
 
 .portfolio {
@@ -100,8 +101,8 @@ button[aria-disabled='true'] {
     grid-column: span 1;
   }
 
-  .scrollable {
-    min-height: 14vh;
+  .ornament {
+    bottom: 8rem;
   }
 }
